Create upload_automox_agent_logfile
Quick way to retrieve a log file of an automox enabled system without physical or netwerk access. This worklet will upload the logfile to a azure blob storage container using a sas-token.

Only fill out the url variable before the slash and sas token variable and you are good to go!